server/requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Database
pymongo==4.9.0

# AI/ML Libraries
google-genai>=0.3.0
textblob==0.17.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0